Improvements to sequence length computation in LSTM models (#859)
* Compute sequence lengths to avoid effects of padding.

* LSTM models compute sequence lengths for the dynamic_rnn functions. Assumes
  sequences end at the first 0.
* Unidirectional LSTM encoders reverse the input sequences using
  tf.reverse_sequence(), in order to take padding into account.
* Bidirectional LSTM encoders don't reverse the input sequences, since it makes
  no sense.
* Use LSTMCell instead of BasicLSTMCell. Oddly enough, this seems to give a
  consistent improvement.

* Fixed variable name.

* lstm_seq2seq_internal_bid_encoder() uses sequence lengths.

* Added a unit test for find_sequence_lengths().

* Convert lengths to integers.

+++ tensor2tensor/layers/common_layers.py
@@ -1300,7 +1300,7 @@ def length_from_embedding(emb):
   Returns:
     a Tensor with shape [batch].
   """
-  return tf.reduce_sum(mask_from_embedding(emb), [1, 2, 3])
+  return tf.cast(tf.reduce_sum(mask_from_embedding(emb), [1, 2, 3]), tf.int32)
 
 
 def mask_leq(target_length, source_length):
